requirements.txt
numpy>=1.21.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pandas>=1.5.0
yfinance>=0.2.18
newsapi-python>=0.2.7
python-dotenv>=0.19.0
scikit-learn>=1.0.2
tqdm>=4.65.0
openai>=1.0.0 